increased precision

+++ PhysObject.h
@@ -23,22 +23,22 @@ public:
 private:
 	Model model;
 
-	glm::float32 mass;
-	glm::vec3 com;
+	glm::float64 mass;
+	glm::f64vec3 com;
 
 	glm::vec3 velocity;
 	glm::vec3 acceleration;
 
-	glm::vec3 omega;
-	glm::vec3 angularAcceleration;
+	glm::f64vec3 omega;
+	glm::f64vec3 angularAcceleration;
 
-	glm::vec3 L;	//angular momentum
+	glm::f64vec3 L;	//angular momentum
 
-	glm::mat3 I;
-	glm::mat3 I_inv;
+	glm::f64mat3 I;
+	glm::f64mat3 I_inv;
 
-	void compute_inertia_tensor(glm::float32 density, glm::mat3* I0, glm::vec3* com, glm::float32* mass, glm::float32* Vol_out );
-	glm::float32 ComputeInertiaMoment(glm::vec3* P, int I);
-	glm::float32 ComputeInertiaProduct(glm::vec3* P, int I, int J);
+	void compute_inertia_tensor(glm::float64 density, glm::f64mat3* I0, glm::f64vec3* com, glm::float64* mass, glm::float64* Vol_out );
+	glm::float64 ComputeInertiaMoment(glm::f64vec3* P, int I);
+	glm::float64 ComputeInertiaProduct(glm::f64vec3* P, int I, int J);
 
 };
